requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.41
pymysql==1.0.2
kafka-python==2.0.2
pymongo==3.13.0